TSK-1937: moved test-api tests to proper package
