+++ lib/taskana-test-api/src/test/java/pro/taskana/testapi/TestPriorityServiceProvider.java
@@ -1,4 +1,4 @@
-package pro.taskana.testapi.tests;
+package pro.taskana.testapi;
 
 import java.time.Duration;
 import java.time.Instant;
