+++ lib/taskana-test-api/src/test/java/pro/taskana/testapi/builder/TaskAttachmentBuilderTest.java
@@ -1,4 +1,4 @@
-package pro.taskana.testapi.tests.builder;
+package pro.taskana.testapi.builder;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
@@ -12,11 +12,9 @@ import pro.taskana.task.api.TaskService;
 import pro.taskana.task.api.models.Attachment;
 import pro.taskana.task.api.models.ObjectReference;
 import pro.taskana.task.internal.models.AttachmentImpl;
+import pro.taskana.testapi.DefaultTestEntities;
 import pro.taskana.testapi.TaskanaInject;
 import pro.taskana.testapi.TaskanaIntegrationTest;
-import pro.taskana.testapi.builder.ClassificationBuilder;
-import pro.taskana.testapi.builder.TaskAttachmentBuilder;
-import pro.taskana.testapi.tests.DefaultTestEntities;
 
 @TaskanaIntegrationTest
 class TaskAttachmentBuilderTest {
