+++ lib/taskana-test-api/src/test/java/pro/taskana/testapi/builder/TaskCommentBuilderTest.java
@@ -1,13 +1,13 @@
-package pro.taskana.testapi.tests.builder;
+package pro.taskana.testapi.builder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static pro.taskana.common.internal.util.CheckedSupplier.wrap;
+import static pro.taskana.testapi.DefaultTestEntities.defaultTestObjectReference;
+import static pro.taskana.testapi.DefaultTestEntities.defaultTestWorkbasket;
 import static pro.taskana.testapi.builder.ClassificationBuilder.newClassification;
 import static pro.taskana.testapi.builder.TaskCommentBuilder.newTaskComment;
 import static pro.taskana.testapi.builder.WorkbasketAccessItemBuilder.newWorkbasketAccessItem;
-import static pro.taskana.testapi.tests.DefaultTestEntities.defaultTestObjectReference;
-import static pro.taskana.testapi.tests.DefaultTestEntities.defaultTestWorkbasket;
 
 import java.time.Instant;
 import java.util.List;
@@ -31,8 +31,6 @@ import pro.taskana.task.api.models.TaskComment;
 import pro.taskana.task.internal.models.TaskCommentImpl;
 import pro.taskana.testapi.TaskanaInject;
 import pro.taskana.testapi.TaskanaIntegrationTest;
-import pro.taskana.testapi.builder.TaskBuilder;
-import pro.taskana.testapi.builder.TaskCommentBuilder;
 import pro.taskana.testapi.security.WithAccessId;
 import pro.taskana.workbasket.api.WorkbasketPermission;
 import pro.taskana.workbasket.api.WorkbasketService;
