+++ qa/intellij/taskana_formatter_intellij.xml
@@ -27,7 +27,6 @@
   </JSCodeStyleSettings>
   <JavaCodeStyleSettings>
     <option name="REPLACE_INSTANCEOF_AND_CAST" value="true" />
-    <option name="CLASS_NAMES_IN_JAVADOC" value="3" />
     <option name="INSERT_INNER_CLASS_IMPORTS" value="true" />
     <option name="CLASS_COUNT_TO_USE_IMPORT_ON_DEMAND" value="999" />
     <option name="NAMES_COUNT_TO_USE_IMPORT_ON_DEMAND" value="999" />
